back/requirements.txt
flask
Flask-Migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-sqlalchemy
Flask-Script
python-dateutil
